requirements.txt
dj-database-url==0.5.0
Django==2.2.8
django-cors-headers==3.1.1
djangorestframework==3.10.3
feedparser==5.2.1
gunicorn==20.0.2
libthumbor==1.3.2
psycopg2==2.8.4
pycrypto==2.6.1
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
